Add examples/repo/scripts-dir-*
For configuring repo's script's directory. Most of these examples
are meant to be copied over and tuned for a specific releases. Look for
"REPLACE_ME" after copying the example and replace as needed.

Move index.yml internal_repos to an array of objects, so that
cf-networking-releae and silk-release's pipline will continue to
work. Also added an example/index.yml when creating a new index.yml for
a new repo.

Added helper ytt function for parsing internal_repos and when
configuring tests so that the pipeline loop is more generic and
re-usable across pielines.