Implement initial wats config
To start, I used a union of the integration configs found in
tas-runtime/test-config-templates/integration-config-gdn-windows.json
and
garden-windows-ci/tasks/generate-integration-config/run

It's probably worth revisiting some of the choices in these configs. For
example, the fact that most tests are actually skipped. Also, the two
configs disagreed on running security group tests.