Run all CATS tests that don't require special configuration
Many of these test suites are not run by default, but should be tested in every pipeline.
More importantly, we were skipping tests that are directly impacted by the releases under test.
For example, we were skipping service discovery tests in the cf-networking pipeline :facepalm:.

We will likely need to add more parameters to the CATs config generation for pipelines that need
to test special configurations -- for example, TCP routing, Isolation Segments, and H2 routing.
Those configurations will need to be parameterized, because we may not want to run them in every
pipeline.

Signed-off-by: David Sabeti <sabetid@vmware.com>