Add shared task for invoking YTT
This task is useful for rendering the PR pipeline manifests for our jobs
that automatically create PR pipelines.

This used to live here: https://github.com/pivotal/tas-runtime/blob/main/ci/tasks/shared/ytt-render.yml

But we need to use inputs from both the wg-app-platform-runtime-ci and
tas-runtime repos, so we are bringing this over like so.

Signed-off-by: Renee Chu <reneec@vmware.com>