export-release: get release version from BOSH director when manifest has 'latest'
This likely only applies to the cf-networking pipeline. Because that pipeline tests both cf-networking and silk,
some of the environment setup is different from other pipelines. The result is that the BOSH manifest lists the
releases versions as 'latest' instead of using the actual dev number.

The first attempt will continue to search for release version in the manfniest. But if that doesn't return
an actual version number, we'll try again and use the BOSH director as the source of truth.